+++ insts/std/CLAR/CLAR.cpp
@@ -17,7 +17,7 @@ CLAR::CLAR() : Instrument()
 	// future setup here?
 }
 
-int CLAR::init(float p[], short n_args)
+int CLAR::init(float p[], int n_args)
 {
 // p0 = start; p1 = dur; p2 = noise amp; p3 = length1; p4 = length2
 // p5 = output amp; p6 = d2 gain; p7 = stereo spread (0-1) <optional>
