+++ insts/std/STRUM/BEND.cpp
@@ -18,7 +18,7 @@ BEND::BEND() : Instrument()
 	// future setup here?
 }
 
-int BEND::init(float p[], short n_args)
+int BEND::init(float p[], int n_args)
 {
 // p0 = start; p1 = dur; p2 = pitch0 (oct.pc); p3 = pitch1 (oct.pc);
 // p4 = gliss function; p5 = fundamental decay time; p6 = nyquist decay time;
