+++ insts/std/STRUM/BEND1.cpp
@@ -22,7 +22,7 @@ BEND1::BEND1() : Instrument()
 	// future setup here?
 }
 
-int BEND1::init(float p[], short n_args)
+int BEND1::init(float p[], int n_args)
 {
 // p0 = start; p1 = dur; p2 = pitch0 (oct.pc); p3 = pitch1 (oct.pc)
 // p4 = gliss function #; p5 = fundamental decay time
