+++ insts/std/TRANSBEND/TRANSBEND.cpp
@@ -66,7 +66,7 @@ TRANSBEND :: ~TRANSBEND()
 }
 
 
-int TRANSBEND :: init(float p[], short n_args)
+int TRANSBEND :: init(float p[], int n_args)
 {
    float outskip, inskip, dur, transp, interval = 0, total_indur, dur_to_read;
    float averageInc;
